esg-as-fcn/requirements.txt
azure-functions
numpy==1.18.4
pandas==1.0.4
scipy==1.4.1
